allowed only creator to put and delete endpoint for event

+++ web_api_packt/planner/routes/events.py
@@ -32,6 +32,7 @@ async def retrive_event(id: PydanticObjectId) -> Event:
 
 @event_router.post('/new')
 async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         'message': 'Event created successfully'
@@ -41,6 +42,18 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
 
 @event_router.put('/{id}', response_model=Event)
 async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if not event:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail='event does not exist'
+        )
+    
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='update not allowed'
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -53,12 +66,19 @@ async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depe
 
 @event_router.delete('/{id}')
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
-    event = await event_database.delete(id)
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail='event does not exist'
         )
+    
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='delete not allowed'
+        )
+    event = await event_database.delete(id)
     return {'message': 'event deleted successfully'}
